+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Changelog
 
   - `renBCH (renBCH) <https://www.coingecko.com/en/coins/renbch>`__
   - `renZEC (renZEC) <https://www.coingecko.com/en/coins/renzec>`__
+  - `Swerve.fi DAI/USDC/USDT/TUSD (swUSD) <https://www.coingecko.com/en/coins/swusd>`__
 
 * :release:`1.9.0 <2020-11-20>`
 * :feature:`717` Uniswap v2 LP balances are now detected by Rotki. Faster balance queries, swaps history and LP events history is also supported for premium users. Finally uniswap trades are now taken into account in the profit/loss report for premium users.